operator: orchestration for rbac, registry, service
This adds the main MiniMummi reconcile logic, where we first
create the headless service for the cluster, followed by
rbac stuff (Role, RoleBinding, and ServiceAccount for it),
and note we use those instead of Cluster<N> because we only
should need them on the level of the namespace. Then we create
the registry, which is a StatefulSet. We use a statefulset
instead of deployment for the possibility of stateful things -
namely persistent volume claims (eventually for artifacts) and
a peristent network name. Next I need to add the rabbitmq
deployment, the mlserver and wfmanager. For rabbitmq people
normally use an operator, but I think that might be overkill
and am going to try doing basic certificate generation
to config maps here first.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>