ensure we use nrpoc for cganalysis/createsims
The job cores that are set was being used for the
scripts, which is problematic because we have to use
the same value for scheduling and ask less. Since we
need to re-run these with ARM and more iterations (10)
I want to fix this and allow both analyses to use the
full machines resources.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>